Raw transport config props proxy.
