+++ packages/ui/src/services/config-transport/legacy/withConfigTransport.spec.disabled.ts
@@ -3,9 +3,9 @@ import { Options, Vue, setup, prop, mixins } from 'vue-class-component'
 import { mount, config } from '@vue/test-utils'
 
 import withConfigTransport, { getComponentOptions, resolveProps } from './withConfigTransport'
-import { GlobalConfig, useGlobalConfig } from '../global-config/global-config'
-import { GlobalConfigPlugin } from '../global-config/global-config-plugin'
-import VaConfig from '../../components/va-config/VaConfig'
+import { GlobalConfig, useGlobalConfig } from '../../global-config/global-config'
+import { GlobalConfigPlugin } from '../../global-config/global-config-plugin'
+import VaConfig from '../../../components/va-config/VaConfig'
 
 const initialConfig = {
   all: { value: 'valueFromContext' },
